Fixed photo display, Responsiveness, Formatting and P5
FINALLY fixed the photo display issue, the problem was the property
100% which was allowing the browser to size the DOM element BEFORE the
photo was loaded, making it seem like nothing was there. This is
finally fixed. Made a TON of other edits here too, added responsiveness
for iPhone 6 (TODO: Responsiveness for template pages for the Design
Process pages)

Made a bunch of formatting changes, and added stuff for landing page
p5.js interactivity. Disabled for now until I come up with a cool idea,
but the functionality is there.